Visual Compass: Basic refactoring
