+++ bitbots_visual_compass/cfg/visualCompass.cfg
@@ -24,10 +24,10 @@ group_compass_akaze = group_compass.add_group("Akaze")
 group_compass_orb = group_compass.add_group("Orb")
 group_compass_sift = group_compass.add_group("Sift")
 
-compass_matcher_enum = gen.enum([gen.const("akaze", str_t, "akaze", "Akaze matcher"),
-                                 gen.const("orb", str_t, "orb", "Orb matcher"),
-                                 gen.const("sift", str_t, "sift", "SIFT matcher")],
-                                "Choose from these matcher types")
+compass_feature_detector_enum = gen.enum([gen.const("akaze", str_t, "akaze", "Akaze feature_detector"),
+                                 gen.const("orb", str_t, "orb", "Orb feature_detector"),
+                                 gen.const("sift", str_t, "sift", "SIFT feature_detector")],
+                                "Choose from these feature_detector types")
 
 group_ros.add("img_msg_topic", str_t, 0, "Subscriber of the image input message", None)
 group_ros.add("img_msg_queue_size", int_t, 0, "Subscriber of the image input message", min=1, max=20)
@@ -39,7 +39,7 @@ group_setup.add("feature_map_file_path", str_t, 0, "Path to the map file in the 
 group_setup.add("feature_map_field", str_t, 0, "Field on which the map was recorded", None)
 
 group_compass_settings.add("compass_type", str_t, 0, "Compass type", "binary", edit_method=compass_type_enum)
-group_compass_settings.add("compass_matcher", str_t, 0, "Matcher algorithm", "sift", edit_method=compass_matcher_enum)
+group_compass_settings.add("compass_feature_detector", str_t, 0, "FeatureDetector algorithm", "sift", edit_method=compass_feature_detector_enum)
 
 group_compass_multiple.add("compass_multiple_map_image_count", int_t, 0, "Count of images to compute the map", min=2, max=360)
 group_compass_multiple.add("compass_multiple_mean_feature_count", int_t, 0, "Big compass_multiple_mean_feature_count -> small confidence for few found features", min=2, max=360)
