+++ bitbots_visual_compass/src/compass/feature_detector/interface.py
@@ -1,0 +1,22 @@
+#!/usr/bin/env python3
+
+class FeatureDetector():
+
+    def __init__(self, config):
+      pass
+
+    # filters kp1, so that only those elements remain, whose corresponding elements
+    # in desc1 match with any element in desc2
+    # kp1 and desc1 must have the same length.
+    def match(self, kp1, desc1, desc2):
+        pass
+
+    def get_keypoints(self, image):
+        pass
+
+    def debug_keypoints(self, image):
+        return image
+
+    def set_config(self, config):
+        pass
+
